UI updation and edit note

+++ app/src/main/java/com/example/todonote/MainActivity.kt
@@ -4,7 +4,9 @@ import android.content.Intent
 import android.os.Bundle
 import android.util.Log
 import android.widget.Button
+import android.widget.Toast
 import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
@@ -26,6 +28,7 @@ class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
+//        finish()
         setContentView(R.layout.activity_main)
 
 //
@@ -40,7 +43,11 @@ class MainActivity : AppCompatActivity() {
 
 
         recyclerAdaptor = TaskAdaptor(
-            { position: Int -> delVal(position)
+            {
+                position: Int -> updateValue(position)
+            },
+            {
+                position: Int ->  delVal(position)
             },
             this,
             taskList
@@ -48,6 +55,7 @@ class MainActivity : AppCompatActivity() {
 
 
         recyclerView.adapter = recyclerAdaptor
+
 
 
         val button : Button = findViewById(R.id.navigate_add_task)
@@ -58,10 +66,38 @@ class MainActivity : AppCompatActivity() {
 
     }
 
+    fun updateValue(position: Int){
+        val taskId = taskList[position].id
+        val taskModel = taskList.get(position)
+        val intent : Intent = Intent(this,AddNewTaskActivity()::class.java)
+
+        intent.putExtra("taskId",taskModel.id)
+        intent.putExtra("taskTitle",taskModel.title)
+        intent.putExtra("taskBody",taskModel.task)
+
+        startActivity(intent)
+    }
+
     fun delVal(position : Int)  {
         val taskId = taskList[position].id
-        taskList.removeAt(position)
-        databaseHandler.deleteTheTask(taskId)
-        recyclerAdaptor.notifyItemRemoved(position)
+        val taskModel = taskList.get(position)
+
+        val builder = AlertDialog.Builder(this)
+
+        builder.setTitle("Delete Note")
+            .setMessage("Do you really delete the note ? \n ${taskModel.title} \n ${taskModel.task.to(20)}")
+            .setPositiveButton("Yes"){dialog,which ->
+                taskList.removeAt(position)
+                databaseHandler.deleteTheTask(taskId)
+                recyclerAdaptor.notifyItemRemoved(position)
+                Toast.makeText(this, "Task deleted", Toast.LENGTH_SHORT).show()
+                dialog.dismiss()
+            }
+            .setNegativeButton("No"){dialog,which->
+                dialog.dismiss()
+            }
+
+        val alertDialog : AlertDialog = builder.create()
+        alertDialog.show()
     }
 }
